% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{count_k_mers_fast}
\alias{count_k_mers_fast}
\title{Count kmers}
\usage{
count_k_mers_fast(string, k)
}
\arguments{
\item{string}{\code{character} with the sequence}

\item{k}{\code{integer} defines length of the subsequence to be counted}
}
\description{
Count kmers
}
\examples{
count_k_mers_fast("AAPGAGAYY", 2)
}
