% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_kmers_str}
\alias{count_kmers_str}
\title{Count k-mers for string sequences}
\usage{
count_kmers_str(s, d, alphabet, pos)
}
\arguments{
\item{s}{a \code{string} vector representing an input sequence}

\item{d}{an \code{integer} vector representing gaps between consecutive elements of k-mer}

\item{alphabet}{a \code{string} vector representing valid elements of k-mer}

\item{pos}{a \code{logical} value that denotes whether positional k-mers should be generated}
}
\value{
a named vector with countings of k-mers
}
\description{
Count k-mers for string sequences
}
\details{
K-mers that contain elements from \code{alphabet} but do not exist in the input sequence are also generated.
}
\examples{
count_kmers_str(
c("a", "b", "c", "d", "x", "y", "z", "z", "a", "a"),
d=c(0,0),
c("a", "b", "c", "z"),
pos=FALSE)
}
