% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{generate_random_alphabet}
\alias{generate_random_alphabet}
\title{Generating random alphabet of given size}
\usage{
generate_random_alphabet(nsize = 10)
}
\arguments{
\item{nsize}{\code{integer} number of elements in the alphabet}
}
\description{
Generating random alphabet of given size
}
