% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu.R
\name{otu}
\alias{otu}
\title{Cluster sequences into operational taxonomic units.}
\usage{
otu(x, k = 5, threshold = 0.97, ...)
}
\arguments{
\item{x}{a "DNAbin" object.}

\item{k}{integer giving the k-mer size used to generate the input matrix
for k-means clustering.}

\item{threshold}{numeric between 0 and 1 giving the OTU similarity cutoff.}

\item{...}{further arguments to be passed to \code{kmeans}.}
}
\value{
an integer vector of cluster membership with values from 1 to
  the total number of OTUs.
}
\description{
This function performs divisive heirarchical clustering on a set of
  DNA sequences using sequential k-means partitioning,
  returning an integer vector of OTU membership.
}
\details{
TBA
}
\examples{
  ##TBA

}
\references{
TBA
}
\author{
Shaun Wilkinson
}
